+++ packages/graphyne-server/src/http/index.ts
@@ -1,0 +1,1 @@
+export { createHandler } from './handler';
