+++ packages/graphyne-server/src/index.ts
@@ -1,2 +1,2 @@
-export { createHandler } from './createHandler';
+export { createHandler as httpHandler } from './http';
 export { GraphyneCore as Graphyne } from 'graphyne-core';
